Use uint16_t with coords

+++ common/includes/Paquet/PaquetPlayerCoord.hh
@@ -9,8 +9,8 @@ class		PaquetPlayerCoord : public Paquet
 private:
   uint8_t	_id;
   uint8_t _playerID;
-  uint8_t _x;
-  uint8_t _y;
+  uint16_t _x;
+  uint16_t _y;
 
 public:
   PaquetPlayerCoord();
@@ -20,11 +20,11 @@ public:
   virtual ~PaquetPlayerCoord();
 
   void setPlayerID(const uint8_t&);
-  void setPosition(const uint8_t&, const uint8_t&);
+  void setPosition(const uint16_t&, const uint16_t&);
 
   const uint8_t& getPlayerID() const;
-  const uint8_t& getX() const;
-  const uint8_t& getY() const;
+  const uint16_t& getX() const;
+  const uint16_t& getY() const;
 
   void		createPaquet();
   void		parsePaquet();
