+++ common/includes/Paquet/PaquetPlayerShot.hh
@@ -10,8 +10,8 @@ private:
   uint8_t	_id;
   uint8_t _playerID;
   uint8_t _type;
-  uint8_t _x;
-  uint8_t _y;
+  uint16_t _x;
+  uint16_t _y;
 
 public:
   PaquetPlayerShot();
@@ -22,12 +22,12 @@ public:
 
   void setType(const uint8_t&);
   void setPlayerID(const uint8_t&);
-  void setPosition(const uint8_t&, const uint8_t&);
+  void setPosition(const uint16_t&, const uint16_t&);
 
   const uint8_t& getType() const;
   const uint8_t& getPlayerID() const;
-  const uint8_t& getX() const;
-  const uint8_t& getY() const;
+  const uint16_t& getX() const;
+  const uint16_t& getY() const;
 
   void		createPaquet();
   void		parsePaquet();
